+++ CoreGraphNode/src/au/gov/asd/tac/constellation/graph/node/plugins/DefaultPluginEnvironment.java
@@ -38,6 +38,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import org.openide.NotifyDescriptor;
 import org.openide.util.Exceptions;
 import org.openide.util.lookup.ServiceProvider;
 
@@ -71,7 +72,7 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
             // A typical use-case is an arrangement followed by a camera reset: obviously doing the reset before the
             // vertices have been relocated is not sensible.
             if (async != null) {
-                for (Future<?> future : async) {
+                for (final Future<?> future : async) {
                     if (future != null) {
                         try {
                             future.get();
@@ -97,7 +98,7 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
 
             try {
                 ConstellationLogger.getDefault().pluginStarted(plugin, parameters, graph);
-            } catch (Exception ex) {
+            } catch (final Exception ex) {
             }
 
             PluginManager manager = new PluginManager(DefaultPluginEnvironment.this, plugin, graph, interactive, synchronizer);
@@ -135,21 +136,22 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
                         }
                     }
                 }
-            } catch (InterruptedException ex) {
+            } catch (final InterruptedException ex) {
                 auditPluginError(plugin, ex);
                 interaction.notify(PluginNotificationLevel.INFO, "Plugin Cancelled: " + plugin.getName());
                 Thread.currentThread().interrupt();
                 if (currentReport != null) {
                     currentReport.setError(ex);
                 }
-            } catch (PluginException ex) {
+            } catch (final PluginException ex) {
                 auditPluginError(plugin, ex);
                 interaction.notify(ex.getNotificationLevel(), ex.getMessage());
                 ex.printStackTrace();
                 if (currentReport != null) {
                     currentReport.setError(ex);
                 }
-            } catch (Exception ex) {
+            } catch (final Exception ex) {
+                NotifyDescriptor.Exception exceptionDialog = new NotifyDescriptor.Exception(ex);
                 auditPluginError(plugin, ex);
                 final String msg = String.format("Unexpected non-plugin exception caught in %s.executePluginLater();%n", DefaultPluginEnvironment.class.getName());
                 LOGGER.log(Level.WARNING, msg, ex);
@@ -165,7 +167,7 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
 
                 try {
                     ConstellationLogger.getDefault().pluginStopped(plugin, parameters);
-                } catch (Exception ex) {
+                } catch (final Exception ex) {
                 }
             }
 
@@ -201,7 +203,7 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
 
         try {
             ConstellationLogger.getDefault().pluginStarted(plugin, parameters, graph);
-        } catch (Exception ex) {
+        } catch (final Exception ex) {
         }
 
         try {
@@ -219,7 +221,8 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
                 plugin.run(graphs, interaction, parameters);
 
             }
-        } catch (Exception ex) {
+        } catch (final Exception ex) {
+            NotifyDescriptor.Exception exceptionDialog = new NotifyDescriptor.Exception(ex);
             auditPluginError(plugin, ex);
             if (currentReport != null) {
                 currentReport.setError(ex);
@@ -237,7 +240,7 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
 
             try {
                 ConstellationLogger.getDefault().pluginStopped(plugin, parameters);
-            } catch (Exception ex) {
+            } catch (final Exception ex) {
             }
         }
 
@@ -272,7 +275,7 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
 
         try {
             ConstellationLogger.getDefault().pluginStarted(plugin, parameters, GraphNode.getGraph(graph != null ? graph.getId() : null));
-        } catch (Exception ex) {
+        } catch (final Exception ex) {
         }
 
         try {
@@ -280,7 +283,8 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
             final PluginInteraction interaction = new DefaultPluginInteraction(manager, currentReport);
 
             plugin.run(graph, interaction, parameters);
-        } catch (Exception ex) {
+        } catch (final Exception ex) {
+            NotifyDescriptor.Exception exceptionDialog = new NotifyDescriptor.Exception(ex);
             auditPluginError(plugin, ex);
             if (currentReport != null) {
                 currentReport.setError(ex);
@@ -298,7 +302,7 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
 
             try {
                 ConstellationLogger.getDefault().pluginStopped(plugin, parameters);
-            } catch (Exception ex) {
+            } catch (final Exception ex) {
             }
 
         }
@@ -333,7 +337,7 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
 
         try {
             ConstellationLogger.getDefault().pluginStarted(plugin, parameters, GraphNode.getGraph(graph != null ? graph.getId() : null));
-        } catch (Exception ex) {
+        } catch (final Exception ex) {
         }
 
         try {
@@ -342,7 +346,8 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
 
             plugin.run(graph, interaction, parameters);
 
-        } catch (Exception ex) {
+        } catch (final Exception ex) {
+            NotifyDescriptor.Exception exceptionDialog = new NotifyDescriptor.Exception(ex);
             auditPluginError(plugin, ex);
             if (currentReport != null) {
                 currentReport.setError(ex);
@@ -366,10 +371,10 @@ public class DefaultPluginEnvironment extends PluginEnvironment {
         return null;
     }
 
-    private void auditPluginError(Plugin plugin, Throwable error) {
+    private void auditPluginError(final Plugin plugin, final Throwable error) {
         try {
             ConstellationLogger.getDefault().pluginError(plugin, error);
-        } catch (Exception ex) {
+        } catch (final Exception ex) {
         }
     }
 }